Create arg_parser.py script
